refactor(cardinal): make ComponentFilter use component.Component (#533)
Closes: WORLD-707

<!---
Add a prefix to indicate what kind of release this pull request corresponds to:
  feat, fix, docs, style, refactor, perf, test, build, ci, chore, revert
--->

## Overview

This is PR 1 out of 2 to refactor the Search system. This PR is focused on refactoring
`filter.ComponentFilter` to work wih `component.Component` instead of `component.ComponentMetadata`.

By doing this, we will be able to make `filter.ComponentFilter` the main way to do filter throughout
the codebase. 

## Brief Changelog

- Refactor ComponentFilter to take in component.ComponentMetadata
- Minor adjustment to CQL to accomodate this new change
- Minor adjustment to tests to accomodate this new change


## Testing and Verifying

- Existing unit tests have been adjusted to accomodate this change

<!---
Pick one of the following options:

- This change is a trivial rework/code cleanup without any test coverage.

- This change is already covered by existing tests, such as <describe test>.

- This change added tests and can be verified as follows:
    - Added unit test that validates ...
    - Added integration tests for end-to-end deployment with ...
    - Extended integration test for ...
    - Manually verified the change by ...
--->