refactor(cardinal): make NewSearch use filter.ComponentFilter and replace LazySearch (#534)
Closes: WORLD-703

<!---
Add a prefix to indicate what kind of release this pull request corresponds to:
  feat, fix, docs, style, refactor, perf, test, build, ci, chore, revert
--->

## Overview

> Description of the overall background and high-level changes that this PR introduces

This is PR 2 out of 2 and completes the Search refactor. 

Since we have refactored `filter.ComponentFilter` to accept `component.Component`, there
is no longer a need for Filterable (filter.go). Therefore, this PR eliminates it altogether.

Search now directly uses `filter.ComponentFilter`, which allows us to eliminate `ConvertToComponentFilter`
call that was previously returning an error. As a result, we can drop the LazyContainer altogether.

There is now a clean **single** combo of filter/search. 

One note here: a side-effect of using ConvertToComponentFilter call is that it validates whether a component
is registered with the `Engine` before doing archetype matching. This is a feature that we might want to preserve
but it is technically not required because `evaluateSearch()` (ecs/search.go#107) will just treat filter with unregistered
component like any other filter (it will just not find a match, etc). If we want this validation, we can create a ticket for this check
to be added in `evaluateSearch` prior to running `sm.SearchFrom`.

## Brief Changelog

<!---
Example:
- The metadata is stored in the blob store on job creation time as a persistent artifact
- Deployments RPC transmits only the blob storage reference
- Daemons retrieve the RPC data from the blob cache
--->

- Remove filterable and adjust `cardinal/filter.go` to use `filter.ComponentFilter`
- Remove LazyContainer/LazyUpdate
- `cardinal.NewSearch` now just uses a vanilla `ecs.NewSearch`
- Replace all uses of Filterable with `filter.ComponentFilter`
- Replace all usage of search such as within `/evm` and `/ecs/persona.go` to use the
  the new search. Eliminating error returns from NewSearch both from internal use and external use. 
- Adjust tests to accomodate the changes. 

## Testing and Verifying

<!---
Pick one of the following options:

- This change is a trivial rework/code cleanup without any test coverage.

- This change is already covered by existing tests, such as <describe test>.

- This change added tests and can be verified as follows:
    - Added unit test that validates ...
    - Added integration tests for end-to-end deployment with ...
    - Extended integration test for ...
    - Manually verified the change by ...
--->

- Unit tests are updated to accommodate the changes